pyproject.toml
[tool.poetry]
name = "zyenyo"
version = "0.1.0"
description = "A Small Bot Made by Me, For fun!"
authors = ["Seizoxu"]

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.5.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
